requirements.txt
aiogram==3.7.0
openai
tavily-python
requests
pymupdf
python-docx
groq
